requirements.txt
streamlit
supabase
gspread
google-auth
pandas
openai
streamlit-option-menu
python-dotenv
plotly
